chore: move `base_settings.py` in its own module
