+++ connectors-sdk/connectors_sdk/exceptions/__init__.py
@@ -1,11 +1,1 @@
 """Offer Exception handling tools to develop connectors."""
-
-from .error import (
-    ConfigError,
-    ConfigValidationError,
-)
-
-__all__ = [
-    "ConfigError",
-    "ConfigValidationError",
-]
