+++ connectors-sdk/connectors_sdk/exceptions/error.py
@@ -1,16 +1,6 @@
 """Offers a collection of custom exceptions to develop connectors."""
 
 
-class ConfigError(Exception):
-    """Base class for configuration-related errors.
-
-    This exception is raised when there is an issue with the configuration of the connector.
-    It can be used to indicate problems such as missing required fields, invalid values, or
-    other configuration-related issues that prevent the connector from starting correctly.
-    It signals a actionable problem in configuration and allows the user to respond appropriately immediately.
-    """
-
-
 class DataRetrievalError(Exception):
     """Base class for data retrieval-related errors.
 
@@ -30,13 +20,4 @@ class UseCaseError(Exception):
     This error can be recoverable or skipped in the application layer,
     especially if it is designed to iterate over entities.
     It signals a problem in the connector's logic or implementation that needs to be addressed.
-    """
-
-
-class ConfigValidationError(ConfigError):
-    """Base class for configuration validation-related errors.
-
-    This exception is raised when there is an issue with the configuration validation of the connector.
-    It can be used to indicate validation problems on the configuration that prevent
-    the connector from starting or working correctly.
     """
