+++ connectors-sdk/connectors_sdk/settings/__init__.py
