+++ connectors-sdk/tests/test_settings/conftest.py
